scripts/requirements.txt
google-auth
requests
